refactor: change package structure_SplashActivity.kt and SplashViewModel
