+++ app/src/main/java/campus/tech/kakao/map/view/SplashActivity.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.view
 
 import android.content.Intent
 import android.os.Bundle
@@ -6,8 +6,8 @@ import android.util.Log
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.Observer
+import campus.tech.kakao.map.viewmodel.SplashViewModel
 import campus.tech.kakao.map.databinding.ActivitySplashBinding
-import campus.tech.kakao.map.view.MapActivity
 import com.google.firebase.remoteconfig.FirebaseRemoteConfig
 import com.google.firebase.remoteconfig.FirebaseRemoteConfigSettings
