+++ app/src/main/java/campus/tech/kakao/map/viewmodel/SplashViewModel.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.viewmodel
 
 import android.app.Application
 import android.util.Log
