+++ app/src/main/res/layout/activity_splash.xml
@@ -7,7 +7,7 @@
     <data>
         <variable
             name="viewModel"
-            type="campus.tech.kakao.map.SplashViewModel" />
+            type="campus.tech.kakao.map.viewmodel.SplashViewModel" />
         <import type="android.view.View"/>
     </data>
